moneymong-220 refactor: 마이몽 package 구조 변경
