+++ feature/mymong/src/main/java/com/moneymong/moneymong/feature/mymong/main/component/MyMongTopBar.kt
@@ -1,4 +1,4 @@
-package com.moneymong.moneymong.feature.mymong.component
+package com.moneymong.moneymong.feature.mymong.main.component
 
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.Text
