+++ feature/mymong/src/main/java/com/moneymong/moneymong/feature/mymong/navigation/MyMongNavigation.kt
@@ -5,7 +5,7 @@ import androidx.compose.foundation.layout.padding
 import androidx.compose.ui.Modifier
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
-import com.moneymong.moneymong.feature.mymong.MyMongScreen
+import com.moneymong.moneymong.feature.mymong.main.MyMongScreen
 
 const val mymongRoute = "mymong_route"
