+++ feature/mymong/src/main/java/com/moneymong/moneymong/feature/mymong/withdrawal/component/WithdrawalTopBar.kt
@@ -1,4 +1,4 @@
-package com.moneymong.moneymong.feature.mymong.component
+package com.moneymong.moneymong.feature.mymong.withdrawal.component
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
